RFC 98: Change emphasis in the summary; clarify "model social norms".
Changing the order of items in the summary was suggested in GitHub discussion. Two people on GitHub
also expressed confusion about the meaning of "model social norms", so we added a sentence in "rules
vs principles" clarifying it. The paragraph that that sentence was added to was getting quite long,
so we split it, but no other wording was changed.