Add tables of alternatives discussed in this RFC
An overview of alternatives and their drawbacks (including drawbacks to
the syntax currently proposed) are provided in tables in the
Alternatives section, as requested by the Nix team.

These tables are intended to be an overview of some of the discussion in
this RFC and will be updated as discussion continues.